feat(sched): add priority-inheriting mutex and fix event flag clearing
- Add syn_mutex primitive with priority inheritance and protothread macros
- Fix PT_BLOCK_EVENT to selectively clear only matched flags in task wait_mask
- Upgrade scheduler rr_per_prio array to uint16_t with bounds validation
- Add Voice Activity Detection (syn_vad), TPM 2.0 (syn_tpm2), Robot Kinematics (syn_kinematics), and MQTT-SN (syn_mqttsn) modules
- Add unit tests achieving 100.00% Line and Function coverage across 308 files

+++ src/syntropic/sched/syn_sched.h
@@ -59,7 +59,7 @@ extern "C" {
 typedef struct SYN_Sched {
     SYN_Task *tasks;   /**< Pointer to caller-owned task array    */
     size_t task_count; /**< Number of tasks in the array          */
-    uint8_t rr_per_prio[SYN_SCHED_PRIO_LEVELS];
+    uint16_t rr_per_prio[SYN_SCHED_PRIO_LEVELS];
     /**< Per-priority round-robin indices      */
 } SYN_Sched;
